Add pytest plugin
Move the betamax fixtures to an installable pytest plugin, so
that dependent applications can also make use of them.
